+++ scm-webapp/src/main/resources/sonia/scm/cli/i18n.properties
@@ -107,17 +107,17 @@ scm.repo.set-role.name = User or group name to set role for
 scm.repo.set-role.role = Permission role to grant to user or group
 scm.repo.set-role.forGroup = Set role for a group, not for a user
 
-scm.repo.add-permission.usage.description.0 = Add single permission for user or group
-scm.repo.add-permission.repository = Repository namespace/name
-scm.repo.add-permission.name = User or group name to grant permission to
-scm.repo.add-permission.verbs = Single permissions to grant to user or group
-scm.repo.add-permission.forGroup = Add permission for a group, not for a user
+scm.repo.add-permissions.usage.description.0 = Add single permission for user or group
+scm.repo.add-permissions.repository = Repository namespace/name
+scm.repo.add-permissions.name = User or group name to grant permission to
+scm.repo.add-permissions.verbs = Single permissions to grant to user or group
+scm.repo.add-permissions.forGroup = Add permission for a group, not for a user
 
-scm.repo.remove-permission.usage.description.0 = Revoke single permission from user or group
-scm.repo.remove-permission.repository = Repository namespace/name
-scm.repo.remove-permission.name = User or group name to revoke permission from
-scm.repo.remove-permission.verbs = Single permissions to revoke from user or group
-scm.repo.remove-permission.forGroup = Revoke permission from a group, not for a user
+scm.repo.remove-permissions.usage.description.0 = Revoke single permission from user or group
+scm.repo.remove-permissions.repository = Repository namespace/name
+scm.repo.remove-permissions.name = User or group name to revoke permission from
+scm.repo.remove-permissions.verbs = Single permissions to revoke from user or group
+scm.repo.remove-permissions.forGroup = Revoke permission from a group, not for a user
 
 scm.repo.clear-permission.usage.description.0 = Revoke all permissions from user or group
 scm.repo.clear-permission.repository = Repository namespace/name
