+++ scm-webapp/src/main/resources/sonia/scm/cli/i18n_de.properties
@@ -117,17 +117,17 @@ scm.repo.set-role.name = Name des Benutzers oder der Gruppe für die Berechtigung
 scm.repo.set-role.role = Rolle, die für den Benutzer oder die Gruppe gesetzt werden soll
 scm.repo.set-role.forGroup = Setzt die Berechtigung für eine Gruppe, nicht für einen Benutzer
 
-scm.repo.add-permission.usage.description.0 = Fügt eine einzelne Berechtigung für einen Benutzer oder eine Gruppe hinzu
-scm.repo.add-permission.repository = Repository Namespace/Name
-scm.repo.add-permission.name = Name des Benutzers oder der Gruppe für die Berechtigung
-scm.repo.add-permission.verbs = Einzelne Berechtigungen, die für den Benutzer oder die Gruppe hinzugefügt werden sollen
-scm.repo.add-permission.forGroup = Setzt die Berechtigung für eine Gruppe, nicht für einen Benutzer
+scm.repo.add-permissions.usage.description.0 = Fügt eine einzelne Berechtigung für einen Benutzer oder eine Gruppe hinzu
+scm.repo.add-permissions.repository = Repository Namespace/Name
+scm.repo.add-permissions.name = Name des Benutzers oder der Gruppe für die Berechtigung
+scm.repo.add-permissions.verbs = Einzelne Berechtigungen, die für den Benutzer oder die Gruppe hinzugefügt werden sollen
+scm.repo.add-permissions.forGroup = Setzt die Berechtigung für eine Gruppe, nicht für einen Benutzer
 
-scm.repo.remove-permission.usage.description.0 = Entzieht einem Benutzer oder einer Gruppe eine einzelne Berechtigung
-scm.repo.remove-permission.repository = Repository Namespace/Name
-scm.repo.remove-permission.name = Name des Benutzers oder der Gruppe, dem/der die Berechtigung entzogen werden soll
-scm.repo.remove-permission.verbs = Einzelne Berechtigungen, die dem Benutzer oder der Gruppe entzogen werden sollen
-scm.repo.remove-permission.forGroup = Entziehe die Berechtigung einer Gruppe, nicht einem Benutzer
+scm.repo.remove-permissions.usage.description.0 = Entzieht einem Benutzer oder einer Gruppe eine einzelne Berechtigung
+scm.repo.remove-permissions.repository = Repository Namespace/Name
+scm.repo.remove-permissions.name = Name des Benutzers oder der Gruppe, dem/der die Berechtigung entzogen werden soll
+scm.repo.remove-permissions.verbs = Einzelne Berechtigungen, die dem Benutzer oder der Gruppe entzogen werden sollen
+scm.repo.remove-permissions.forGroup = Entziehe die Berechtigung einer Gruppe, nicht einem Benutzer
 
 scm.repo.clear-permission.usage.description.0 = Entzieht einem Benutzer oder einer Gruppe alle Berechtigungen
 scm.repo.clear-permission.repository = Repository Namespace/Name
